Implement init_signal_handlers
